New Fuctionality
- Validate auth
- Need to be auth for edit and delete contacts and appointments

+++ backend/routes/contacts.js
@@ -2,10 +2,11 @@ const express = require('express');
 const router = express.Router();
 
 const contact = require('../models/contact');
+const {verifyToken} = require('./validationFunctions');
 
 //Get Contacts
-router.get('/getContacts', (req, res) => {
-    contact.find((err, docs) => {
+router.get('/getContacts', verifyToken, (req, res) => {
+    contact.find({createdById: req.userId},(err, docs) => {
         if(err){
             res.status(500).end();
         } else {
@@ -15,7 +16,7 @@ router.get('/getContacts', (req, res) => {
 });
 
 // Add Contact
-router.post('/addContact', (req, res) => {
+router.post('/addContact', verifyToken,  (req, res) => {
     const newContact = new contact(req.body);
     newContact.save((err, doc)=>{
         if(err){
@@ -29,27 +30,35 @@ router.post('/addContact', (req, res) => {
 });
 
 // Edit Contact
-router.post('/editContact/:id', (req, res) => {
+router.post('/editContact/:id', verifyToken, (req, res) => {
     const contactId = req.params.id;
 
     console.log(req.body);
 
-    contact.findByIdAndUpdate(contactId,req.body,(err, doc) => {
+    contact.findOneAndUpdate({_id: contactId, createdById: req.userId}, req.body,(err, doc) => {
         if(err) {
             res.status(500).end();
         }else{
-            res.status(200).end();
+            if(doc === null){
+                res.status(404).end();
+            }else{
+                res.status(200).end();
+            }
         }
     });
 });
 
 // Delete Contact
-router.delete('/deleteContact/:id' , (req, res) => {
-    contact.findOneAndRemove(req.params.id, (err, doc) => {
+router.delete('/deleteContact/:id', verifyToken, (req, res) => {
+    contact.findOneAndRemove({_id:req.params.id, createdById:req.userId}, (err, doc) => {
         if(err){
             res.status(500).end();
         }else{
-            res.status(200).end();
+            if(doc === null) {
+                res.status(404).end();
+            }else{
+                res.status(200).end();
+            }
         }
     });
 });
